restructure api
