+++ api/src/admin/controller.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-console */
 
+/** ADMIN CONTROLLER */
+
 /** Node Imports */
 import { Request, Response } from 'express';
 import Mongoose from 'mongoose';
