+++ api/src/locations/controller.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-console */
 
+/** LOCATION CONTROLLER */
+
 import { Request, Response } from 'express';
 import Mongoose, { MongooseDocument } from 'mongoose';
