+++ api/src/search/controller.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-console */
 
+/** SEARCH CONTROLLER */
+
 /** Node Imports */
 import Mongoose from 'mongoose';
 import { Request, Response } from 'express';
